[WAL-5] Added Checkout Config Provider view model – Checkout block now obsolete and removed
